LayoutTests/platform/ios-sim-deprecated/css3/device-adapt/opera/constrain-018-expected.txt


FAIL CSS Test: @viewport constrained - Small non-auto height not affecting width. assert_equals: expected 1 but got 600(stack: assert@file:///Volumes/Data/WebKit.git/LayoutTests/resources/testharness.js:2168:73
assert_equals@file:///Volumes/Data/WebKit.git/LayoutTests/resources/testharness.js:705:15
file:///Volumes/Data/WebKit.git/LayoutTests/css3/device-adapt/opera/constrain-018.xhtml:69:22
step@file:///Volumes/Data/WebKit.git/LayoutTests/resources/testharness.js:1186:30
onload@file:///Volumes/Data/WebKit.git/LayoutTests/css3/device-adapt/opera/constrain-018.xhtml:68:14)
